social-media-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "social-media-agent"
version = "0.1.0"
description = "AI-powered social media content automation agent for influencers"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "openai>=1.50.0",
    "tweepy>=4.14.0",
    "requests>=2.31.0",
    "pillow>=10.2.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.4",
    "sqlalchemy>=2.0.25",
    "streamlit>=1.40.0",
    "plotly>=5.18.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[project.scripts]
social-agent = "social_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/social_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
